Changed aproach to using Selenium Wait
The script now waits until the target game is clickable using WebDriverWait from Selenium.

If there's no carousel block, the game should be clickable from the start, provided that there is no cookie banner.

If the wait does time out, the script tries to close the cookie banner again, just to be sure.