Script can click through carousel again
If the script can't click on a game, it will try to click on a carousel control button. If it can't, it waits around 60 seconds so that the carousel might cycle itself to the desired item.

If after 60 seconds the script still can't click on  the game, it attempts to remove the cookie bar again, and tries to click one more time.

If this last try still didn't work the script skips to the next game in the list.

Added back user_agent because it seems I removed that by accident